fix metrics field injection into AOP-advised beans
When injecting metrics into annotated fields of AOP proxied beans, the
bean post processor was attempting to inject the values into the proxy
instance, not the underlying bean.  Since the fields in question didn't
exist, this would cause an error.  Instead, perform the injection at an
earlier phase of bean post-processing, which targets the underlying
bean, not the proxy wrapper.